fix(database_writer): Remove unique columns for keywords table
If a test is calling the same keyword in a test, we would get an error
informing us we violated the unique constraint.

A keyword should not be limited by any unique constraint since each
test can call it multiple times, with multiple args, .. and all these
scenarios have the same name and type which will result in a unique
constraint violation.